+++ NEWS.md
@@ -1,8 +1,10 @@
 # tidytransit (dev)
 - Empty strings in txt files are always parsed as `NA` in `read_gtfs()` [#229](https://github.com/r-transit/tidytransit/pull/229)
 - `raptor()` and `travel_times()` now support in-seat transfers [#225](https://github.com/r-transit/tidytransit/pull/225)
-- Updated documentation authenticated downloads [#224](https://github.com/r-transit/tidytransit/pull/224)
+- Updated documentation on authenticated downloads [#224](https://github.com/r-transit/tidytransit/pull/224)
+- The parameter `max_departure_time` in `travel_times()` is no longer supported
 - Improved error handling
+
 
 # tidytransit 1.7.1
 - update gtfs reference sysdata [#221](https://github.com/r-transit/tidytransit/pull/221)